refactor: app Providers の QueryClient/trpcClient 生成を lib/trpc へ抽出
isAuthError/handleAuthError と QueryClient 生成を lib/trpc/query-client.ts
へ、trpcClient(httpBatchLink)生成を lib/trpc/browser-client.ts へロジック
不変のまま移動し、Providers.tsx は Provider 合成のみに縮小する。

Refs #1532
